requirements.txt
Django>=4.1.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
requests==2.28.2